hide hitbox

+++ jeu/src/main/java/fr/ul/acl/model/Monstre.java
@@ -66,7 +66,7 @@ public class Monstre extends Entity{
 	}
 
 	public int get_speed() {
-		return this.speed;
+		return (int) this.speed;
 	}
 
 	@Override public int get_level(){
